Increased CSRF protection and support for 2 and 3-column primary keys

+++ handlers/drop.php
@@ -7,6 +7,17 @@ if (! User::require_admin ()) {
 	exit;
 }
 
+$f = new Form ('get', $this);
+if (! $f->verify_csrf ()) {
+	header ('Location: /admin');
+	exit;
+}
+
+if (! preg_match ('/^[a-zA-Z0-9_]+$/', $_GET['table'])) {
+	header ('Location: /admin');
+	exit;
+}
+
 if (! db_execute ('drop table `' . $_GET['table'] . '`')) {
 	$page->title = i18n_get ('Error');
 	printf ("<p><a href='/dbman/index'>&laquo; %s</a></p>\n", i18n_get ('Back'));
@@ -14,7 +25,7 @@ if (! db_execute ('drop table `' . $_GET['table'] . '`')) {
 	return;
 }
 
-$page->title = i18n_get ('Table Dropped') . ': ' . $_GET['table'];
-printf ("<p><a href='/dbman/index'>&laquo; %s</a></p>\n", i18n_get ('Back'));
+$this->add_notification (i18n_get ('Table Dropped') . ': ' . $_GET['table']);
+$this->redirect ('/dbman/index');
 
 ?>
